add policy to category and sub_category

+++ app/controllers/api/v2/admin/categories_controller.rb
@@ -1,16 +1,18 @@
 class Api::V2::Admin::CategoriesController < Api::V2::Admin::BaseController
   def index
-    categories = Category.all
+    categories = policy_scope [:admin, Category]
     render json: categories.order(:id), status: 200
   end
 
   def show
     category = Category.find(params[:id])
+    authorize [:admin, category]
     render json: category, status: 200
   end
 
   def create
     category = Category.new(category_params)
+    authorize [:admin, category]
 
     if category.save
       render json: category, status: 201
@@ -21,6 +23,7 @@ class Api::V2::Admin::CategoriesController < Api::V2::Admin::BaseController
 
   def update
     category = Category.find(params[:id])
+    authorize [:admin, category]
 
     if category.update(category_params)
       render json: category, status: 200
@@ -31,6 +34,7 @@ class Api::V2::Admin::CategoriesController < Api::V2::Admin::BaseController
 
   def destroy
     category = Category.find(params[:id])
+    authorize [:admin, category]
     category.destroy
     head 204
   end
@@ -38,6 +42,6 @@ class Api::V2::Admin::CategoriesController < Api::V2::Admin::BaseController
   private
 
   def category_params
-    params.permit(:name)
+    params.permit(policy([:admin, Category]).permitted_attributes)
   end
 end
